+++ changelog.md
@@ -1,3 +1,12 @@
+# Version 2025.10.10 (2025-10-10)
+
+- Add file_path to BasePersitentFile load
+- Add get_latest_response_by_params to session recorder
+- Add 'optional settings' config option
+- Remove individual ttl for session recorder entry
+- Rename filename to file_name
+- Use enum for internal custom ids
+
 # Version 2025.10.9 (2025-10-18)
 
 - Add session recorder